Fix updating contacts
Moved the the updating directly to the promise in the person updater
function, as it's probably easiest that way without the use of async
stuff, and it still allows the logging. Also, forgot that can't
have curly brackets in arrow functions without the use of return,
caused undefineds to be returned.