Add "Full Stack Phonebook"
Made the changes to the frontend's backend interface such that a
relative address is used, and changes the backend such that the frontend
build can be used.

I'm not sure how I should actually be doing this, though, as now
the frontend application can't really be run separately. I guess
you can always choose an earlier commit and test it there, but that's
quite cumbersome.