Add "Blog List Expansions" steps 7-8 (4.19-4.20)
Functionality seems to be as expected and tests seem to generally
run okay. Though, there is some weirdness going on with using "before" in
the top scope, and occasionally, I'm getting an "ECONNRESET" with
a status code of -4077, which ends up failing a single test case from
time to time. Would need to look at the cause a little further ideally,
but the actual implementation seems to otherwise be correct.