Fix ECONNRESET issue
At least hopefully fixed the issue with the ECONNRESET (see 5c20f88).
The expectation was that the weirdness with the "before" was related
to the error, and moving the "before" to the relevant describe block from
the top scope seems to have fixed the issue. It seems to suggest
that perhaps the "before" was not able to always run properly on the
top scope, and the connection with the database was not established
properly.