Add "Blog List Expansions" step 10 (4.22)
Currently the middleware can send an error response if the
id is invalid or does not exist or whatever. While this works with
what it's currently used for, one may also wish to have a differently
rendered webpage depending on whether authorization goes through,
for example, something like looking at your own profile and being
able to change it vs. looking at someone else's profile. Irrelevant
as such, but it may be more useful to simply let the router's routes
handle the parsing of the user id themselves, and this mostly makes the middleware
rather pointless: already the verification of the token might throw
and error, which may not be what's needed in a given route.

In turn, could just rename the middleware to something more specific
than userExtractor, I guess, but that's what it's supposed to be called
for this exercise.