Add functionality related to "Blog List Frontend" step 2 (5.2)
The issue is that on the frontend, I'm supposed to save a token from
a login. However, there is no way before this commit to really
fetch a user or any of their details from the backend using a token
based on the exercises I've done thus far, so need to add something
of the sort.