Add "Blog List Frontend" step 4 (5.4)
Added better timeout stuff this time so it doesn't cancel out back-to-back
notifications. In turn, it's still possible that a notification during
login and then a quick successful login could result in the notification
being still displayed once on the logged-in screen, which is slightly
un-neat if nothing else. But it's not like it's a polished app anyway.