Fix "Blog List Frontend" step 5 (5.5)
Scrolled a little further down and realised the blog submission form
is also supposed to close after submission, so added that. I did do
it in a way that seems like it might be a little sketchy for how React
is supposed to be built, though it does at least seem to work: the app
gives the form a callback (this was already the case from before), which
the form calls with the blog details (which could also be passed as just
an object, really). The callback attempts to save the blog, and either
returns true or false depending on success. The return is awaited by
the form, and the state is set based on the return. So, the communication
is child -> parent -> child, but execution sort of weirdly happens
in the child throughout. I guess React wants parents
to pass data through props mostly, while children call callbacks to "return"
state back, in a sense.

The issue is that returning the success of the blog adding through props
seems a little convoluted, even if it would be the correct way, and
I'm not too happy about having all kinds of barely necessary, sort of cryptic
states hanging around in the parent. I can't immediately find anything
that disallows the use of this way of doing it, either. It's clean in one
sense, but maybe breaks the consistency a little.

The issue in using this method, though, is that BlogAddForm needs to know
about how the callback function is formed. In turn, technically the
BlogAddForm is in charge of what is passes to the function, and the parent
is actually the one that needs to adhere to the BlogAddForm's requirements.
Only issue then is that now, the callback needs to specifically return
a promise with a boolean value, which may be a little strict for the partly
unspecified nature of the callback function.

In other news, the blog form currently retains the contents written in it,
regardless of whether adding a new blog was successful or not. I'm currently
not sure whether I want to delete the contents after a successful addition,
as one may wish to reuse some of the values, and it's not that difficult
to get rid of them. (Is it better to have a "clear all" and leave the inputs
as they are? or is this akin to the discouraged use of the "reset" functionality?)