Add "Blog List Frontend" step 11 (5.11)
Had to also work on the backend a little, as the blog http methods
did not all return the blog in the form that the GET does, main
difference being the lack of a populated user property. Does take
an extra database call, I guess, but I would say it makes most sense
that if the methods do return the blog, then the blog should be in the
same state as if fetched by a GET.

Might have some similar problems elsewhere, but I'll look into that
later.