Add "Anecdotes" step 8 (6.10)
Kind of getting the feeling that there's a tad too much of magic
that JS devs want to use. Using the createSlice, the actions
returned don't allow for code inspection to see what values
they might take, just the action and payload arguments. Trying
the "prepare" approach, This problem is not solved. Also "No,
I don't want to write try...catch, I'll use a library that does
that automatically."

In general, the amount of magic happening does make it a little
difficult to follow how it all gets constructed. Immer magically
allows direct mutation of objects (or mutation syntax to be used);
function names get magically transformed into action type strings
and such; useDispatch refers to the implicitly present store object,
defined on some higher level in another file, and passing an action
to it calls the reducer created in a third file where the property
of an object (or the name of the createSlice object?) and the inspected
names of the functions (or inner objects) in the "reducers" property
of the createSlice (which is imported from a fourth file) define the string
name of the action taking place, and the function defined in the createSlice
defines both the action creator and the reducer at the same time, or something like that
(I assume, I haven't really taken a close look at the whole thing).
Granted, it's not written in an easy-to-read manner,
but good lord is that complex to follow. Maybe
it'll make sense someday if I start writing massive branching codebases,
but it certainly seems like a wacky approach right now.

I just find myself asking often "wait, so where does THIS thing get defined now?",
and the answer is usually "behind the scenes", or so it seems. And yeah, all that's fine
and dandy once you've learnt it, but looking at the redux library docs
and all the features, it's like you're having to learn an entire
language for the "convenience". Redux itself is presumably a good
idea, but avoiding boilerplate like the plague makes the code fairly
obtuse after a while. Hmm, It's almost like there's an anecdote about
this, or something... Well, maybe the issue is that I'm not nearly
as smart as whoever came up with the Redux Toolkit.