Add "Routed Anecdotes" step 1 (7.1)
Routing should work now, but as I'm (I believe) using the newer method
of writing React Routers, I'm having to move things around quite a
bit. The main issue just now is that I need to actually share the
the state somehow. I guess the react reducer stuff is good for that.