Add "React Query and Context" step 1 (7.1)
Just changed the App to use the reducer instead of the state,
which is in practice not much different. I suppose the idea
with using the Contexts is that you can access them from anywhere
without prop drilling, and that's how you should be using them.
The problem right now is that I would have to refactor basically the entirety
of the application in whole because the notifications are usually
set in some function that gets passed to the children of the App.
As such, I'll just leave it as it is for now, as I assume I'll be
making the change to the use of Contexts for other stuff too, and
once that's done, it'll make more sense to move the management
of the state to the individual components.