Add "React Query and Context" step 3 (7.12)
I'm getting the feeling that the main use of React Query is the ability
to easily share the state between the components. useMutation is nice because
of the ability to separately define the onSuccess and everything, and
I guess there's some other niceties when you dig deeper, but I also
realised that it's not really necessary per-se: I can still just make
a normal backend call which requires far less boiler plate.

Similarly, the state of the application also need not be kept in React Query,
as is evident by the use of the react contexts. I could just as well
create a context for the blogs and just fetch that data at the start,
setting the blogs into the context, otherwise updating only on the client
side when not refreshing outright.

This is fairly obvious as originally we only used React alone to handle
the state. However, the main issue of this approach was having to pass
props deep into the component tree, a problem that can clearly be
solved by just using Context instead of State. Still, something
like Query does offer some extra utilities which may well be useful
eventually, like the concept of staleness, so it's not like it's
completely pointless. It just doesn't feel like it's that useful
for our use case.