Add "Automatic Code Formatting" (7.9)
Completely forgot about this one. Kind of wild how many of the rules
are opposite to what I'd do. "Prettier" is also a funny choice,
I'd certainly argue that semicolons do not make the code prettier
(at least in the aesthetic sense), even though the code may be better
that way. The same goes for some other stuff, notably the double-spaced
indent just makes it seem claustrophobic. "If you're running out of space,
split that stuff into more functions or use better formatting otherwise"
is what they say (not the Prettier people, though), and I agree. And
I would've agreed with the {} formatting for functions a while back,
but then I came around on it looking cleaner the other way, and
now you're saying I need to do it the first way anyway?

Also, Prettier is not the ultimate solution anyway. The way it should
work is that every developer develops using their preferred look (possibly
having to concede in the case of the semicolons, for example), and
the code is formatted to the team's agreed upon format when pushing
or whatever. Of course, I guess this is not that easy to achieve in
practice as VCS usually use diffs for changes. Although, the comparison
can be done if the formatting step is always made before the comparison.
This would require a configuration in the VCS so that the formatter
can be defined. Is that already a thing, I wonder?  (There is some
stuff about pre-commit hooks, but I doubt that does what I'd want it to do.)