Add "Individual User View" (7.15)
It is a little inefficient as it's having to fetch all users.
This is of course due to the decision made on the backend that
necessitates that a valid token matching the user be passed when
fetching an individual user. The point at the time was, I think,
that I needed a way to use the token to get the username, as only
the token was saved, or something, although actually right now the
username is saved too. I don't remember anymore.

In any case, I'll keep the backend as-is for now, because it's
proof-of-concept either way and the backend is at least currently
not part of the development.