Add "Blog View" (7.16)
Feels a little janky, but it seems to work. Does not have any of the
features now on the Blogs view, don't know if that was the idea.

It's quite difficult to manage the state without having to repeat
code and while allowing a parent component to still define the
child's behaviour. This latter part would especially become an
issue if you need to actually do tests, as the default behaviour
cannot necessary be run, or for testing it doesn't make sense to run it.