Add "Adding a book" (8.10)
I'm now thinking what exactly is the point of GraphQL. The immediate
and obvious upside is that you're able to better define what you
want to access on the client side, whereas usual REST only allows
for preset queries. The problem is that if you need to mutate anything,
keeping things up-to-date may be quite difficult if you have different
queries for the same items, e.g. get books with all props vs. just the title.
As I understand it, these would supposedly work as two different queries,
meaning I would have to re-run both in order to get the same data; if
I only query all items, the title query does not update (unless the queries
actually fetch all the data, then just return the asked-for stuff, but
this seems a little odd?). The issue is the same for local updates.
This means that ultimately the most practical approach is to just
query everything (no "just the title" queries) as then it's easy
to keep it all up-to-date.

The alternative is always refetching and still using separate queries,
but this is likely going to be less efficient than using the same fetch-all query
for all relevant cases and only refetching when refreshing the page
completely (and naturally doing local updates in other cases).

The other thing is that the query is more functional, allowing
for searching for only a given genre, title, or author etc. without
a need for specifying these queries in a very specific way on the backend.
Of course, you could perform the necessary work on the client side too
(for filtering in this example), but for a large database, this is not ideal.

So, having to always explicitly specify which properties I want with
the returned items can be a little annoying (although, you could also
specify an extra virtual property or something?) as I feel like I'd
rather always want all the properties. In turn, it might be that a data
point has some properties, but maybe
I don't want all the properties ultimately on the client side, in which
case it is nice to be able to only fetch some props,
but I still don't understand not having a SELECT *.
Having the ability to modularly change the query on the client side is
overall not too bad.