Add "Authors birth year", "Authors birth year advanced" (8.11-8.12)
Just used the select box immediately.

I read up on the whole normalisation thing in Apollo Client and realised
that that is what what makes the client somewhat more useful. Had to
change the authors to use "name" as key, though in practice it would
probably be best to just always return the IDs when working with the client,
as that is what the client uses by default for the keys.

With the normalisation, there is no need to update or refetch the updated
author manually, so that's pretty cool.