Add partial "Database" part 2 (8.14)
Oh My Sweet Baby Jesus. How does toObject transform stuff AS ITS BEING
PUT INTO MONGODB? That is not at all the expected behaviour. Or am I
accidentally transforming stuff to objects and saving in the database?
Does the Schema not control how it's supposed to be in the database?

Point being, for some reason using "flattenIds" or whatever with
toObject meant that the books were being saved to the database with
IDs as strings instead of ObjectIds, and for whatever reason mongoose
will not "find()" IDs unless they're ObjectIds in the database (regardless
of whether you use the ObjectId or string)? Absolutely ridiculous.
In any case, looks like the book query works as expected now.