Add "Books by genre, part 1", "books by genre with GraphQL" (8.19, 8.21)
Not really sure whether this would be the idea with the filtering
using GraphQL. It doesn't really specify and I'd imagine the point
is to just use the genre argument in the book query. Still, I need
to find the set of genres using JS as there's no query setup for
just getting the queries on the backend, though obviously such a
query would be easy to add.

Besides utilising GraphQL features, I'm also not sure whether it's
really any better to do it like this, seeing as it's just as easy (computation-wise)
to filter the list in JS directly, and requires no backend communication
after the first time. This way does currently ensure that all the genres
are most definitely accounted for, though. The overall situation could
naturally change as more data needs to be fetched from the backend,
though I guess you'd usually fetch data in batches, having pages of
data or otherwise incrementally loading in more data.