Change to using localStorage directly for login management (8.)
I realised that reducers are supposed to be pure, and the way I'd
written the login "reducers" are not. The bottom line is that I want
to have the login state as a context, and I want to be potentially
able to control it from a number of locations. The actually obvious
choice was ultimately to just use the localStorage directly, though
this does slightly change the way it works, namely not being to
have a globally updatable login state variable as that state now
has to be explicitly queried instead. Although, I could still have
the usual useState context, and just use setState and the currently
written localStorage queries together to, well, set the state.
Doing it like this is practically the same as using reducers
(setState vs. dispatch), except reducers require purity while useState
does not care how the state is set.