Add "Up-to-date cache and book recommendations" (8.22)
Bit of a two-birds situation as the use of '""' as the genre argument
in fetching the books meant that the no-arguments version was never
fetched and could not be updated, which I apparently only today
realised caused issues with adding books.

I had been quite curious
about the whole refetch thing as it does indeed seem to work such
that only the specifically defined query is fetched, which can
often be a little difficult. With the books' genre filter defined the
way it is, the refetch would need to be written for each individual
genre, and this is just wacky. cache.evict seems to do what I've actually
wanted, which is to effectively remove all the queries of a certain
type, regardless of argument. Much like with writing every query separately,
updating the cache of every individual query would be quite complex to do,
especially if more than just the genre is used as argument.
Causing a total refetch of all the queries of a certain type is really
the correct solution, because it's the least soul-crushingly complex option
while still being better than a complete reload of the page.

As a side note,
it's interesting that we're now considering efficiency in this, while
back in the day when internet speeds and such would have been lower,
we just always reloaded (though maybe there was less to send too, to an
extent). Even now many web pages work on a reload basis, opening
a whole new page every time. I guess the optimisations are still a great help on the
server side, and you can argue about climate change and such too.