Add "n+1" (8.26)
No-op as doing the query only does one mongoose operation.
I guess the assumption was that I would not have the
book IDs saved with the Author, but I did, since way back when. If I
didn't, I'd have to look through the books to find the matching
author, which would cause a whole lot of queries.