Add "Command line" (9.3)
Does seem to work okay when the arguments are passed correctly,
but without using the " -- ", the run command thinks negative arguments
are actually flags to the command. In any case, using the " -- " is correct,
I believe, as it's used to pass arguments to the underlying, actually
executed command.

I'm doing a slightly hacky thing with the export to get around the
script vs. module problem in TypeScript, and as I don't want to
steer to much away from how I've been instructed to write the files.