Add "Exercise 9.16" (9.16)
Bit of an ad hoc styling on the whole thing.

While this certainly teaches about how one might use the literals
for this kind trick, could I not also simply use conditional
rendering, or indeed conditionality in general? The idea of
defining a separate type for each slightly different object
seems wacky. Seems a lot easier to just have some properties
be optional then check for their existence, would be a lot easier
with adhering to DRY this way.