Add partial "Exercise 9.17", functionality but no proper rendering (9.17)
Not including the backend as it is not part of the exercise directly.
I'm using the https://github.com/fullstack-hy2020/flight-diary/tree/part3
version, the one which uses zod and actually has cors setup unlike
the main version. I'm also currently using some of the main one's
typing utils, for now anyway, as that's what I had first before I
realised that there was no cors in the backend and decided to change
to the branch that actually has cors.