Add "Exercise 9.19" (9.18)
I used validation in the frontend already, so it's not very
easy to test backend errors just now. In any case, I reckon it's
proof of concept of an error handling method, what I've implemented.

Funny to realise that reducers are not actually needed either for global
state,
though they may be a recommended way to deal with global state.
The only issue is that they assume purity, so it's a little iffy
to use them if you want impure functions. In turn, I don't know
if I should have impurity anyway.