Add "Patientor, step 6" (9.26)
I'm still thinking that it seems
Kind of wacky to use separate entry types as it does make defining
the component a lot more difficult. Would be a lot more concise
if I could just conditionally render values, the conditional being
whether the property is present or not. Of course the point is
about having rigorous code, which this might be better for,
but anyway.