Add "Patientor, step 7" (9.27)
A little bit of redundancy in parsing the NewEntry, but in practice
the route should probably take care of making sure that the request
body is correct, something that may not be needed in all cases (e.g. when
just calling the add function locally as opposed to (effectively) by HTTP request proxy).
At the same time, you'd still want the function that actually
adds the stuff to ensure that it is added in the correct format,
seeing as that's going to be necessary to do every time anything
is added anyway.