Add "Patientor, step 8" (9.28)
This time I did also add some better error comms between the backend
and frontend, but mostly to avoid the potentially awful Zod dumps
from the backend. They shouldn't happen anyway as the parsing's already
being done on the frontend, but still.

I would argue that the frontend should not really
know any specific details of why the backend failed, mostly just
a 500 error. The backend should show what went wrong, but that is
not the user's business. Still, it's perhaps nice to visually let the user
know that something went bang in the backend.