Add partial "Patientor, step 9", setting up Occup. Entry (9.29)
Pained over trying to figure out a nice modular setup which would
allow changing only the added entry stuff while keeping the base
entry data when creating a new form (i.e. choosing a new entry type
in the middle of the process without losing all inputted data), and a setup
that would allow the least amount of unnecessary repetition. Decided
to just leave such considerations for later and just get a working
setup first. Not that I'm actually going to necessary do any refactoring,
but I do want to try and write sensible code regardless of whether
I plan on doing anything with it in the future.