Change to more sensible requirements for forms (9.)
No sense, for example, in requiring that every hospital entry necessarily
has a time of discharge as I would imagine that there can be multiple
entries relating to a person's hospital visit, and it's not like
the entry would be made only once the patient has already been discharged
anyway. Of course, this sort of makes the hospital entry type kind of
useless in most cases, but it wasn't my decision to add such a type.
In practice, you'd probably eventually add the possibility of
changing an entry.