Add build and compose to backend (12.21)
Not as sure as to how to work with genuine secrets with Docker,
though it seems to me like passing the .env is pretty much
the same as just using through dotenv. There's secrets available
for compose, but these would require some extra fiddling where as
here I'm able to just pass the .env unmodified. And ultimately,
I largely struggle to see how I'd be any less secure like this compared
to just using dotenv like normally. With the secrets, based on what
I've seen, they're still readable in the container as well. Realistically
the most useful thing to do is to just not have any massively important
secrets available during development, so use a separate dev user for
the database, for example.