Add "creating a containerized production setup" (12.22)
Decided to leave most of the testing out of the Dockerfiles as
it's not very doable such as they are. It would be more practical
to have a separate setup for testing, like test.Dockerfile and
a similar compose such that everything can be run through those,
if one wishes to do testing using containers anyway.